refactor: clean up Appearance, ProcessPanel and fix video playback
- Extract AppearanceTab into components/settings/AppearanceTab.tsx (out of 3400-line settings.tsx)
- Add FONT_SIZE_MIN/MAX constants, replace magic numbers
- Replace SlidersHorizontal reset icon with RotateCcw
- Restore JSDoc block in ProcessPanel
- Fix video playback in Electron: convert data: URLs to blob URLs in VideoPreviewModal (CSP blocks data: in <video>), revoke on unmount
